requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cloudinary==1.36.0
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==0.5.0
dj-rest-auth==2.1.9
dj3-cloudinary-storage==0.0.6
Django==3.2.22
django-allauth==0.45.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.12.4
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
platformdirs==3.11.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7

